Added TileManager to rooms

+++ hvn3/include/hvn3/views/ViewUpdateEventArgs.h
@@ -7,7 +7,7 @@ namespace hvn3 {
 	class ViewUpdateEventArgs : public UpdateEventArgs {
 
 	public:
-		ViewUpdateEventArgs(float dt, Size<int> room_size);
+		ViewUpdateEventArgs(double dt, Size<int> room_size);
 		const Size<int> RoomSize() const;
 
 	private:
